Add support for enable/disable multiple logins with the same OTP
* Why this change is necessary?
OTP Tokens shouldn't be never used again after their first validation,
to prevent security holes. If a TOTP token is allowed to be used
multiple times during it's lifetime, a man-in-the-middle attacker could
use the eavesdropped token at the same time with the legitimate user,
gaining access to the service but without the user nor the
administrator can notice the fact.
An OTP already used should be rejected, and the user alerted
with an informative message saying to wait for the next token before
entering again. The service administrator could log all the multiple
attempts, and point out those access with the same token but
coming from different IP.

* How does it address the problem?
It adds a new item named :allow_multiple_otp_logins to the
configuration file. When its value is false (default), the database
is checked every time the user try to login with a valid OTP token.
If the otp token is already present in the database, the token is
rejected and the login will fail. Instead if the valid token was never
used before, it is stored in the database, and the login will succeed.

* Are there any side effects?
The database is updated every time a token is valid and never
used before.

# 50-character subject line
#
# 72-character wrapped longer description. This should answer:
#
# * Why was this change necessary?
# * How does it address the problem?
# * Are there any side effects?
#
# Include a link to the ticket, if any.